Move dependency handling to the right folder and update kumi
