+++ include/eve/arch/abi.hpp
@@ -9,7 +9,7 @@
 
 #include <eve/arch/expected_cardinal.hpp>
 #include <eve/arch/float16.hpp>
-#include <eve/detail/kumi.hpp>
+#include <eve/deps/kumi.hpp>
 #include <eve/arch/abi_of.hpp>
 
 #include <concepts>
