+++ include/eve/arch/cpu/as_register.hpp
@@ -10,7 +10,7 @@
 #include <eve/arch/expected_cardinal.hpp>
 #include <eve/detail/abi.hpp>
 #include <eve/detail/has_abi.hpp>
-#include <eve/detail/kumi.hpp>
+#include <eve/deps/kumi.hpp>
 #include <eve/forward.hpp>
 #include <eve/traits/as_wide.hpp>
 #include <cstring>
