+++ include/eve/arch/fundamental_cardinal.hpp
@@ -11,7 +11,7 @@
 #include <eve/arch/cardinals.hpp>
 #include <eve/arch/spec.hpp>
 #include <eve/detail/meta.hpp>
-#include <eve/detail/kumi.hpp>
+#include <eve/deps/kumi.hpp>
 #include <type_traits>
 #include <limits>
 #include <utility>
