+++ include/eve/module/algo/algo/views/detail/preprocess_zip_range.hpp
@@ -8,7 +8,7 @@
 #pragma once
 
 #include <eve/module/core.hpp>
-#include <eve/detail/kumi.hpp>
+#include <eve/deps/kumi.hpp>
 #include <eve/traits.hpp>
 
 namespace eve::algo::views
