+++ include/eve/module/algo/algo/views/zip_iterator.hpp
@@ -16,7 +16,7 @@
 #include <eve/module/algo/algo/views/detail/preprocess_zip_range.hpp>
 
 
-#include <eve/detail/kumi.hpp>
+#include <eve/deps/kumi.hpp>
 
 #include <array>
 #include <concepts>
