+++ include/eve/module/bessel/detail/kernel_bessel_jy_large.hpp
@@ -7,7 +7,7 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/detail/kumi.hpp>
+#include <eve/deps/kumi.hpp>
 #include <eve/module/bessel/detail/kernel_bessel_jy.hpp>
 #include <eve/module/bessel/detail/kernel_bessel_j0.hpp>
 #include <eve/module/bessel/detail/kernel_bessel_j1.hpp>
