+++ include/eve/module/core/regular/impl/ifrexp.hpp
@@ -10,7 +10,7 @@
 #include <eve/concept/value.hpp>
 #include <eve/detail/apply_over.hpp>
 #include <eve/detail/implementation.hpp>
-#include <eve/detail/kumi.hpp>
+#include <eve/deps/kumi.hpp>
 #include <eve/module/core/constant/exponentmask.hpp>
 #include <eve/module/core/constant/half.hpp>
 #include <eve/module/core/constant/maxexponentm1.hpp>
