+++ include/eve/module/core/regular/impl/store.hpp
@@ -11,7 +11,7 @@
 #include <eve/concept/value.hpp>
 #include <eve/concept/ptr_translation.hpp>
 #include <eve/detail/implementation.hpp>
-#include <eve/detail/kumi.hpp>
+#include <eve/deps/kumi.hpp>
 #include <eve/conditional.hpp>
 #include <eve/memory/aligned_ptr.hpp>
 #include <eve/memory/soa_ptr.hpp>
