+++ include/eve/module/core/regular/zip.hpp
@@ -8,7 +8,7 @@
 #pragma once
 
 #include <eve/arch.hpp>
-#include <eve/detail/kumi.hpp>
+#include <eve/deps/kumi.hpp>
 #include <eve/traits/overload.hpp>
 #include <eve/concept/same_lanes.hpp>
